+++ plugins/postcss-cascade-layers/.tape.mjs
@@ -24,8 +24,11 @@ postcssTape(plugin)({
 	'anon-layer': {
 		message: "supports anonymous layer usage",
 	},
-	'examples/example': {
-		message: "minimal example",
+	'invalid-nested-css': {
+		message: "ignores nested css",
+	},
+	'extensions': {
+		message: "css custom extensions",
 	},
 	'unlayered-styles': {
 		message: 'supports unlayered styles alongside layers',
@@ -40,7 +43,7 @@ postcssTape(plugin)({
 			onConditionalRulesChangingLayerOrder: 'warn',
 			onImportLayerRule: 'warn'
 		},
-		warnings: 3,
+		warnings: 4,
 	},
 	'warnings:with-postcss-import': {
 		message: 'correctly handles warnings when postcss-import is used',
@@ -51,7 +54,11 @@ postcssTape(plugin)({
 		},
 		warnings: 2,
 		plugins: [
-			postcssImport(), /* postcss-import must run first */
+			postcssImport({
+				nameLayer: (index) => {
+					return `anon-layer-import--${index}`;
+				}
+			}), /* postcss-import must run first */
 			plugin(),
 		]
 	},
@@ -63,5 +70,8 @@ postcssTape(plugin)({
 	},
 	'specificity-buckets-b': {
 		message: "creates non overlapping specificity buckets",
+	},
+	'examples/example': {
+		message: "minimal example",
 	},
 });