+++ plugins/postcss-cascade-layers/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Cascade Layers
 
+### Unreleased (major)
+
+- Updated: Support for Node v14+ (major).
+
 ### 1.0.5 (July 8, 2022)
 
 - Fix case insensitive `@layer` matching (`@LaYeR`).