+++ plugins/postcss-cascade-layers/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - Updated: Support for Node v14+ (major).
 
+### Unreleased
+
+- Fix broken `@keyframes` in `@layer`.
+- Add support for `@container` as a conditional rule.
+
 ### 1.0.5 (July 8, 2022)
 
 - Fix case insensitive `@layer` matching (`@LaYeR`).