+++ plugins/postcss-cascade-layers/CHANGELOG.md
@@ -3,7 +3,12 @@
 ### Unreleased (major)
 
 - Run `postcss-cascade-layers` early compared to other PostCSS plugins (breaking)
-- Fix broken `@keyframes` in `@layer`
+- Updated: Support for Node v14+ (major).
+
+### Unreleased
+
+- Fix broken `@keyframes` in `@layer`.
+- Add support for `@container` as a conditional rule.
 
 ### 1.0.5 (July 8, 2022)